requirements.txt
flask==3.1.3
ultralytics==8.4.26
huggingface_hub==0.27.1
opencv-python-headless==4.10.0.84
numpy>=1.26.4
python-dotenv==1.0.1
PyYAML==6.0.2
gunicorn==23.0.0

# Testing
pytest==8.3.3
pytest-cov==5.0.0

# Security / CI
pip-audit==2.7.3
